Made do_jump command work
This is a fix(/hack) to make the do_jump command work.
It's probably not the most elegant solution, as so far we only handled nav_commands
and I had to hijack some of the already existing methods,
in particular find_next_nav_index in commands_process.pde.
We probably would like a find_next_index function instead!
But it works reliably and we reset the jump-counter up on re-entry in to auto-mode too.

Signed-off-by: Simon Egli <simon.egli@tik.ee.ethz.ch>